Move the `RulesScriptTemplate` to the Evaluator
Rules are specific to the Evaluator and better fit here.
